all seems to be working

+++ caba-abac.py
@@ -1,12 +1,11 @@
 import subprocess, readline, argparse, sys, os, shutil
 from utils.ruleset import check_your_privilege 
-from utils.utils import login, create_file, get_metadata, set_metadata
-from utils.entities import Subject, Resource
+from utils.utils import login, create_file, get_metadata
+from utils.entities import Subject, Resource, ROLES, DEPARTMENTS, COURSES
 from db.db_tables import load_organization
-from utils.organization import SUBJECTS_LIST, RESOURCE_LIST
-from db.db_interface import add_subject, add_resource, subject_row, resource_row, password_row, print_table
+from db.db_interface import add_subject, resource_row, print_table, delete_subject, delete_resource, delete_password
 
-allowed_cmds = ["ls", "pwd", "echo"]
+allowed_cmds = ["ls", "pwd", "echo", "clear"]
 internal_cmds = ["read", "write", "touch"]
 current_dir = os.path.dirname(os.path.abspath(__file__))
 shell = "caba>"
@@ -62,9 +61,10 @@ def run_shell(args):
                     "whoami                    -> Print your attributes\n" \
                     "read  <filename>          -> Read file\n" \
                     "write <filename>          -> Opens file for writing in nano\n" \
-                    "new-resource <res-name>   -> Create a resource \n" \
-                    "new-subject <new-uid>     -> Create a new subject\n" \
-                    "edit-subject <uid>        -> Edit an existing subject\n" \
+                    "new-resource <res-name>   -> Create a resource (alias: nr)\n" \
+                    "new-subject               -> Create a new subject (alias: ns)\n" \
+                    "delete-subject <sub-id>   -> Delete a subject (alias: ds)\n" \
+                    "db                        -> Inspect the database. (Admin only)\n" \
                     "exit, quit                -> Exits the program\n" \
                     "\nSHELL COMMANDS:\n"\
                     "ls, pwd, echo, touch, rm  -> These commands work like normal"
@@ -82,7 +82,7 @@ def run_shell(args):
                 print("Please provide a path")
 
         # create a resource
-        elif cmd_tokens[0] == "new-resource":
+        elif cmd_tokens[0] in ["new-resource", "nr"]:
             if len(cmd_tokens) > 1:
                 path = cmd_tokens[1].split("/")[-1] # currently only allow files in current dir
                 create_file(path, SUBJECT)
@@ -92,17 +92,21 @@ def run_shell(args):
         # remove files in from current dir
         elif cmd_tokens[0] == "rm":
             if len(cmd_tokens) == 2:
-                path = cmd_tokens[1]
+                path = cmd_tokens[1].split("/")[-1]
                 if os.path.exists(path):
                     id = get_metadata(path)
                     res = resource_row(id)
                     if res != None:
                         READ, WRITE, EXECUTE, OWN = check_your_privilege(SUBJECT, res)
                         if OWN:
-                            try:
-                                os.remove(path)
-                            except Exception as e:
-                                print("Error: Could not remove file")
+                            deleted = delete_resource(id, path)
+                            if deleted:
+                                try:
+                                    os.remove(path)
+                                except Exception as e:
+                                    print(f"Error: Could not remove file at path: {path}")
+                            else:
+                                print(f'Error: no resource in database with name: {path}')
                         else:
                             print("Error: you lack privileges to remove this file")
                     else:
@@ -155,33 +159,99 @@ def run_shell(args):
                 print("Error: No path specified")
 
         # create a new subject
-        elif cmd_tokens[0] in ["new-subject","ns"] :
-            if len(cmd_tokens) == 2:
-                sub_id = cmd_tokens[1]
+        elif cmd_tokens[0] in ["new-subject", "ns"]:
+            if len(cmd_tokens) == 1:
                 if SUBJECT.role == "admin":
+                    id = input('Enter id: ')
+                    name = input('Enter name: ')
+
+                    print(f'Enter role (0-{len(ROLES) - 1}):')
+                    for i, role in enumerate(ROLES):
+                        print(f'{i}. {role}')
+                    role_input = input('> ').strip()
+                    if role_input != '':
+                        role = ROLES[int(role_input)]
+                    else:
+                        role = 'guest'
+
+                    print(f'Enter department (0-{len(DEPARTMENTS)})')
+                    for i, dept in enumerate(DEPARTMENTS):
+                        print(f'{i}. {dept}')
+                    dept_input = input('> ').strip()
+                    if dept_input != '':
+                        departments = {DEPARTMENTS[int(dept_input)]}
+                    else:
+                        departments = {}
+
+                    print(f'Enter subdepartments (0-{len(DEPARTMENTS)}) separated by commas')
+                    for i, dept in enumerate(DEPARTMENTS):
+                        print(f'{i}. {dept}')
+                    nums_string = input('> ').split(',')
+                    int_list = {int(x) for x in nums_string if x.strip().isdigit() and 0 <= int(x) <= len(DEPARTMENTS) - 1}
+                    subdepartments = {DEPARTMENTS[i] for i in int_list}
+
+                    yn = input('Is this the chair of the department (y/n)')
+                    is_chair = True if yn in ['y', 'Y', 'yes'] else False
+
+                    print(f'Enter courses taught (0-{len(COURSES)}) separated by commas')
+                    for i, course in enumerate(COURSES):
+                        print(f'{i}. {course}')
+                    nums_string = input('> ').split(',')
+                    int_list = {int(x) for x in nums_string if x.strip().isdigit() and 0 <= int(x) <= len(COURSES) - 1}
+                    courses_taught = {COURSES[i] for i in int_list}
+
+                    print(f'Enter courses taken (0-{len(COURSES)}) separated by commas')
+                    for i, course in enumerate(COURSES):
+                        print(f'{i}. {course}')
+                    nums_string = input('> ').split(',') 
+                    int_list = {int(x) for x in nums_string if x.strip().isdigit() and 0 <= int(x) <= len(COURSES) - 1}
+                    courses_taken = {COURSES[i] for i in int_list}
+
+                    sub = Subject(
+                        id=id,
+                        name=name,
+                        role=role,
+                        departments=departments,
+                        subdepartments=subdepartments,
+                        courses_taught=courses_taught,
+                        courses_taken=courses_taken
+                    )
                     try:
-                        add_subject(sub_id) 
-                    except: # TODO: raise an error in the db to indicate specifically that the entry already exists and catch it here
+                        add_subject(sub) 
+                    except:
                         print("Error: subject already exists")
                 else:
                     print("Error: you lack privileges to create subjects")
             else:
                 print("Error: no subject specified")
 
-        # edit an existing subject
-        elif cmd_tokens[0] in ["edit-subject", "es"]:
-            print("edit an existing subject")
+        elif cmd_tokens[0] in ['delete-subject', 'ds']:
             if len(cmd_tokens) == 2:
-                sub = cmd_tokens[1]
-                if SUBJECT.role == 'admin':
-                    pass
-                # TODO: database call to edit the subject
+                if SUBJECT.role == "admin":
+                    id = cmd_tokens[1]
+                    yn = input('Are you sure? (yn)')
+                    if yn in ['y', 'Y', 'yes', 'Yes', 'YES', 'YES!']:
+                        delete_subject(id)
+                        delete_password(id)
                 else:
-                    print('Error: cannot modify subjects')
+                    print("Error: you lack privileges to delete subjects")
             else:
-                print("Error: no subject specified")
+                print('Improper arg structure')
+
+        # edit an existing subject
+        # elif cmd_tokens[0] in ["edit-subject", "es"]:
+        #     print("edit an existing subject")
+        #     if len(cmd_tokens) == 2:
+        #         sub = cmd_tokens[1]
+        #         if SUBJECT.role == 'admin':
+        #             pass
+        #         # TODO: database call to edit the subject
+        #         else:
+        #             print('Error: cannot modify subjects')
+        #     else:
+        #         print("Error: no subject specified")
         
-        elif cmd_tokens[0] == "db":
+        elif cmd_tokens[0] == "db" and SUBJECT.role == 'admin':
             print_table('Subjects')
             print_table('Resources')
 
@@ -204,7 +274,6 @@ if __name__=="__main__":
                     description='ABAC policy simulator',
                     epilog='Text at the bottom of help')
     parser.add_argument('-d', '--dev', action='store_true', help='Developer mode')
-    parser.add_argument('-g', '--guest', action='store_true', help='Guest mode')
     parser.add_argument('-rl', '--reload', action='store_true', help='Reload Database')
     args = parser.parse_args()
 
@@ -215,8 +284,6 @@ if __name__=="__main__":
     # check for developer mode
     if args.dev:
         SUBJECT = Subject(id="admin1", role="admin")
-    elif args.guest:
-        SUBJECT = Subject(id='guest', role='guest')
     else: 
         SUBJECT = login()
     run_shell(args)
